pyproject.toml
[project]
name = "vdb-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chonkie[all]>=1.4.0",
    "crawl4ai>=0.7.4",
    "fastmcp>=2.12.4",
    "mcp>=1.17.0",
    "pydantic-ai>=1.0.17",
    "weaviate-client>=4.17.0",
]
